fix: DFRNet training pipeline never actually loaded the SVTR encoder
Bring the DFRNet smoke-train pipeline to a working state:

- fix train.py/dfrnet/model.py param name mismatches (svtr_cfg/num_classes,
  ctc_encoder/ctc_fc attribute names)
- fix build_dataloader() call sites for the installed PaddleOCR's config
  shape (Train/Eval nesting, capitalized mode, required logger, Global
  section for CTCLabelEncode); add configs/digits_dict.txt
- fix paddle API mismatches: randn_like -> randn, MultiHeadAttention return
  value, ctc_loss dtype requirements (label int32, lengths int64)
- fix evaluate()'s post_process/metric call shape

Root cause of the 'loss never decreases' smoke result: _load_pretrained's
checkpoint key remap was missing an '.encoder.' path segment
(head.ctc_encoder.encoder.* in the real checkpoint vs the assumed
head.ctc_encoder.*), so the SVTR encoder was silently left randomly
initialized every run while the loader still reported hundreds of matched
keys (from the backbone). Added tools/baseline_check.py to catch this by
running raw inference against the checkpoint outside the training loop;
confirmed fixed with 20/20 exact matches at 0.9-1.0 confidence.

After the fix, a 15-epoch smoke run on the water-meter dataset shows
loss_main converging (0.06 -> ~0.0001-0.12) and eval acc stabilizing at
~0.92-0.93.

+++ dfrnet/loss.py
@@ -46,8 +46,8 @@ class DFRNetLoss(nn.Layer):
         return F.ctc_loss(
             log_probs,
             labels.cast("int32"),
-            input_lengths.cast("int32"),
-            label_lengths.cast("int32"),
+            input_lengths.cast("int64"),
+            label_lengths.cast("int64"),
             blank=self.blank_idx,
             reduction="mean",
         )
